AI_backend/pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.118.1",
    "fitz>=0.0.1.dev2",
    "llama-cloud-services>=0.6.73",
    "pymupdf>=1.26.5",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "unstructured[pdf]>=0.18.15",
    "python-docx>=1.1.0",
    "nltk>=3.8.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.12.0",
    "openai>=2.8.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
]
